Fix(api): suspend functions for cleaner code
Avoids memory leaks and seamless with viewModelScope.launch usage.
